added logic deleted image from cloud drive

+++ app/src/main/java/com/example/omega/imageviewer/ui/screens/viewer/ImageViewerAdapter.java
@@ -77,6 +77,6 @@ public class ImageViewerAdapter extends BaseRecyclerImageAdapter<BaseRecyclerIma
     public interface OnImageClickListener {
         void onImageClick(long position);
 
-        void onImageLongClick(long position);
+        void onImageLongClick(int position);
     }
 }
